Переделал структуру сайта для следующего урока. Вызов функции для формирования галереи находится в файле pub/html/gallery.html. Сама функция описана в файле engine/functions.php. Сделать модальное окно по инструкции из методики не получилось. Есть подозрения на старую версию jQuery, т.к. я использовал jquery CDN ссылку, а инструкция 2014 года.
