+++ shop/pub/css/style.css
@@ -14,11 +14,15 @@ body {
     margin: 0;
     font-family: "Open Sans", sans-serif;
     color: #5e5e5e;
+    max-height: 100%;
+    height: 100%;
 }
 
 .container {
     margin: 0 auto;
     max-width: 1100px;
+    max-height: 100%;
+    height: 100%;
 }
 
 /* начало оформление header */
