+++ shop/pub/html/footer.html
@@ -11,6 +11,6 @@
 			 ?> Все права защищены</p>
        </div>
     </div>
-    
+
 </body>
 </html>
