+++ shop/pub/html/gallery.html
@@ -1,0 +1,11 @@
+	<div class="container">
+        <h1><?php echo $pageTitle ?></h1>
+        <div class="catalog-row">
+
+<?php
+    echo getImages(GALLERY_DIR);
+?>
+
+
+        </div>
+    </div>
