+++ shop/pub/html/menu.html
@@ -5,6 +5,7 @@
 		<div class="menu">
 			<a href="index.php">Главная</a>
 			<a href="index.php?pageName=catalog">Каталог</a>
+			<a href="index.php?pageName=gallery">Галерея</a>
 			<a href="index.php?pageName=contacts">Контакты</a>
 		</div>
 	</div>
